Completed AI Movement
AI will bounce back and forth from node to node.

The path the AI takes can be assigned from the Unity inspector,
by dragging and dropping PathNodes that the enemy will follow.
Word of caution for now, do not add path nodes that cause the enemy to
fall off edges as the AI will not know to jump, for now.